3-3norm対応
