--- ForSubmit/srcs/3_4_parse_read_map.c
@@ -1,14 +1,0 @@
-/* ************************************************************************** */
-/*                                                                            */
-/*                                                        :::      ::::::::   */
-/*   3_4_parse_read_map.c                               :+:      :+:    :+:   */
-/*                                                    +:+ +:+         +:+     */
-/*   By: yokitaga <yokitaga@student.42tokyo.jp>     +#+  +:+       +#+        */
-/*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/10/20 12:31:24 by yokitaga          #+#    #+#             */
-/*   Updated: 2023/10/20 12:31:29 by yokitaga         ###   ########.fr       */
-/*                                                                            */
-/* ************************************************************************** */
-
-#include "../include/cub3d.h"
-
